+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Use UPNP to determine external IP address ([#3237](https://github.com/hoprnet/hoprnet/pull/3237))
 - Use a relay tag to announce in the DHT that a node acts as a relayer for a specific node ([#3237](https://github.com/hoprnet/hoprnet/pull/3237))
 - Added simulated NAT to E2E tests ([#3237](https://github.com/hoprnet/hoprnet/pull/3237))
+- Automatic deployment of NAT nodes to GCloud ([#3165](https://github.com/hoprnet/hoprnet/issues/3165))
 
 ---